76dollshop-announcement-system/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "authlib>=1.6.6",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.46.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=2.14.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
